final back-up of the project / create readme.md

+++ src/pages/Products/index.jsx
@@ -1,9 +1,21 @@
+/**
+ * Products page 
+ * 
+ * @author: John Dave Lagdameo
+ * @since: 04/13/2023 
+ * 
+ * @internal revisions:
+ * + 
+ * +
+ * +
+ */
 import { useEffect, useState  } from 'react';
-import PropTypes from 'prop-types'
-import Section from '../../components/Section'
-import ProductForm from './ProductForm';
+import Swal from 'sweetalert2';
 
-const Products = props => {
+import ProductForm from './ProductForm';
+import Section from '../../components/Section';
+
+const Products = () => {
 
   const [reload, setReload] = useState(true);
   const [selectedProduct, setSelectedProduct] = useState(null); 
@@ -26,10 +38,28 @@ const Products = props => {
   }
 
   const deleteHandler = (product) => {
-    const newProductList = [...productList].filter(v => v.id !== product.id);
-    setSelectedProduct(newProductList);
-    localStorage.setItem("products", JSON.stringify(newProductList));
-    setReload(true);
+    Swal.fire({
+      title: "Product",
+      text: "Are you sure you want to delete this product?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes'
+    }).then((result) => {
+
+      if (result.isConfirmed) {
+        const newProductList = [...productList].filter(v => v.id !== product.id);
+        setSelectedProduct(newProductList);
+        localStorage.setItem("products", JSON.stringify(newProductList));
+        
+        Swal.fire('Success!', "Product successfully deleted.", 'success');
+        setReload(true);
+        
+      }
+
+    });
+
   }
 
   const onChangeFilter = (v) => {
@@ -40,8 +70,7 @@ const Products = props => {
 
   return (
     <Section>
-      <div className='row'>
-
+      <div className='row mb-4'>
         <div className='col-lg-6'>
           <div className='row'>
             <div className='col-lg-6'>
@@ -55,44 +84,47 @@ const Products = props => {
                 placeholder='Search for keywords...' />
             </div>
           </div>
-          <table 
-            style={{width: "100%" }}
-            className='table table-bordered text-center'>
-              <thead>
-                <tr>
-                  <th style={{width: "35%" }}>Name</th>
-                  <th style={{width: "35%" }}>Price</th>
-                  <th style={{width: "30%" }}>Action</th>
-                </tr>
-              </thead>
-              <tbody>
-                {filteredProductList.map((v) => (
-                   <tr key={v.id}>
-                      <td>{v.name}</td>
-                      <td>{v.price}</td>
-                      <td>
-                        <button 
-                          onClick={() => editHandler(v)}
-                          className='btn btn-sm btn-dark me-1'>
-                          Edit
-                        </button>
-                        <button 
-                          onClick={() => deleteHandler(v)}
-                          className='btn btn-sm btn-outline-dark'>
-                          Delete
-                        </button>
-                      </td>
-                   </tr>
-                ))}
-                {filteredProductList.length === 0 && (
+          <div class="table-responsive mt-3">
+            <table 
+              id="table-product"
+              style={{width: "100%" }}
+              className='table table-bordered text-center'>
+                <thead>
                   <tr>
-                    <td colSpan={3}>
-                      No Product/s to show.
-                    </td>
+                    <th style={{width: "35%" }}>Name</th>
+                    <th style={{width: "35%" }}>Price</th>
+                    <th style={{width: "30%" }}>Action</th>
                   </tr>
-                )}
-              </tbody>
-          </table>
+                </thead>
+                <tbody>
+                  {filteredProductList.map((v) => (
+                    <tr key={v.id}>
+                        <td>{v.name}</td>
+                        <td>{v.price}</td>
+                        <td>
+                          <button 
+                            onClick={() => editHandler(v)}
+                            className='btn btn-sm btn-dark me-1'>
+                            Edit
+                          </button>
+                          <button 
+                            onClick={() => deleteHandler(v)}
+                            className='btn btn-sm btn-outline-dark'>
+                            Delete
+                          </button>
+                        </td>
+                    </tr>
+                  ))}
+                  {filteredProductList.length === 0 && (
+                    <tr>
+                      <td colSpan={3}>
+                        No Product/s to show.
+                      </td>
+                    </tr>
+                  )}
+                </tbody>
+            </table>
+          </div>
         </div>
 
         <div className='col-lg-6'>
@@ -108,6 +140,4 @@ const Products = props => {
   )
 }
 
-Products.propTypes = {}
-
-export default Products
+export default Products;
